move change pass activity to user registration dir, as on previous commit
