+++ HypeChatApp/app/src/main/java/com/fiuba/hypechat_app/activities/user_registration/ChangePasswordActivity.kt
@@ -1,10 +1,11 @@
-package com.fiuba.hypechat_app.activities
+package com.fiuba.hypechat_app.activities.user_registration
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.widget.Toast
 import com.fiuba.hypechat_app.R
+import com.fiuba.hypechat_app.activities.ProfileActivity
 import com.google.firebase.auth.FirebaseAuth
 import kotlinx.android.synthetic.main.activity_change_password.*
