+++ HypeChatApp/app/src/main/res/layout/activity_change_password.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".activities.ChangePasswordActivity"
+        tools:context=".activities.user_registration.ChangePasswordActivity"
         android:background="@drawable/general_background">
